backend/requirements.txt
fastapi
uvicorn
python-multipart
python-docx
mammoth
beautifulsoup4
requests
markdown
python-dotenv
pdfplumber
reportlab
pypdf
urllib3
pytesseract
Pillow
